Add UPC link support for NetBSD
This is sufficient to compile and link a UPC "Hello, World!" program
with clang-upc on NetBSD-6.1.4/amd64.  However, the runtime fails at
startup.

BUILD NOTE:
Additionally, I needed to pass -DLIBUPC_ENABLE_BACKTRACE=FALSE to
cmake due to a lack of backtrace() and backtrace_symbols() in
NetBSD-6.  They are introduced in NetBSD-7.0, but in a separate
libexecinfo which will some require additional logic to link
conditionally.