Add UPC link support for FreeBSD
This is sufficient to compile and link a UPC "Hello, World!" program
with clang-upc on FreeBSD-10.0-RELEASE-p4/amd64.  However, the runtime
fails at startup.

BUILD NOTE:
Additionally, I needed to pass -DLIBUPC_ENABLE_BACKTRACE=FALSE to
cmake.  While backtrace() and backtrace_symbols() are present in
FreeBSD-10.0 (though missing in earlier releases), are in a separate
libexecinfo.  Some additional logic could/should be added to link
with "-lexecinfo" when an a FreeBSD erlease with backtrace support.