Further fix for configure/Makefile build.
(1) LLVM configure adds -Wno-maybe-uninitialized based on the compiler that is used to build LLVM.  However, our library is compiled with just built clang, not the compiler LLVM tested, and Makefile structure uses top level LLVM Makefile defines.

(2) With COMPILE_DEPS settings we make sure include files are installed into the Clang build tree before we compile the source.

(3) Do not build clang-upc-lib.h locally as it contains some macros that are also provided on the command line and compiler will issue a warning.  Build it into the Clang pre-release area which is included as a system include (same as cmake build).