Fix for upc2c issue #78 as manifested on ppc64
This commit corrects the reported alignment for the struct PTS on
a target with 32-bit pointers.  It should be the same as a 32-bit
integer, not a 64-bit one as was the case previously.

This did not show on x86_64, where the ILP32 ABI (x86) requires only
4-byte alignment for both 32- and 64-bit integers.  However on PPC64
both the LP64 and ILP32 ABIs specify 8-byte alignment of 8-byte
integers, leading to a mismatch between clang-upc and the C code in
UPCR.