Correct PTS passing and return on ppc32 target
This commit adds the necessary logic to correctly pass and return
pointers-to-shared on a ppc32 target.

Prior to this commit the PTS was always passed and returned by
reference.  That is correct for -fupc-pts=struct but does not match
the C ABI for -fupc-pts=packed.  So, on a ppc32 target one could use
clang-upc directly (since it was self-consistent) and one could use it
with UPCR with the struct PTS.  However use with UPCR and the packed
PTS would fail due to the difference in calling convention between the
UPC code compiled by clang-upc and the C code in UPCR compiled by
clang (or gcc).