Add support for systems that do not support scalable endpoints.
If system supports scalable endpoints, only one endpoint is created
and scalable endpoints are created for each of the runtime
sub-systems (gmem, lock, ..).  Runtime is used to exchange data on
one point only.

If scalable endpoints are not supported, separate endpoints are
created for each of the sub-systems (with only one scalable endpoint
at the top of it).  Each sub-system exchanges endpoint data on its own.

At this point configure is used to select scalable endpoints.  By
default, support for scalable endpoint is OFF.