Do not use FI_RMA_EVENT for shutdown logic.
In the past we used RAM event capability to signal the remote thread
about shutdown request.  However, as some of the providers do not
support RAM events, a simple spin-loop had to be added (spin-loop is
on the local memory variable written to by the remote thread).

This codes avoids all these by using the tagged send/receive interface
as our shutdown logic is not one-sided communication in nature, there
is always someone to pickup the message - shutdown pthread.  Note that
regular send/recv can be used, however, usage of tagged interface might
allow us to have a separate 'service' EP that we can use across the
run-time.

This implementation can be used as a guidance for implementing similar
things in: locks, barrier, broadcast, collectives.  Locks and broadcast
can use tagged send/recv to completely replace signaling.  However,
barrier and collectives can get better performance if RMA events are
supported.